pyproject.toml
[project]
name = "research_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.9",
    "langchain[anthropic]>=1.2.10",
    "langgraph>=1.0.8",
    "mlflow>=3.10.0",
    "pymupdf",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "wikipedia>=1.4.0",
    "xmltodict>=1.0.4",
]
